+++ src/plsqlDocumentSymbol.provider.ts
@@ -1,9 +1,11 @@
 import * as vscode from 'vscode';
 import PlSqlParser from './plsqlParser.vscode';
+import PLSQLSettings from './plsql.settings';
 
 export class PLSQLDocumentSymbolProvider implements vscode.DocumentSymbolProvider {
 
     public provideDocumentSymbols(document: vscode.TextDocument, token: vscode.CancellationToken): vscode.SymbolInformation[] {
+        PlSqlParser.initParser(PLSQLSettings.getCommentInSymbols());
         return PlSqlParser.getAllSymbols(document);
     }